+++ fishfft/genbun.f90
@@ -302,7 +302,7 @@ SUBROUTINE GENBUN(Nperod,N,Mperod,M,A,B,C,Idimy,Y,Ierror,W)
       CALL POISP2(M,N,W(iwba),W(iwbb),W(iwbc),Y,Idimy,W,W(iwb2),W(iwb3),&
         W(iww1),W(iww2),W(iww3),W(iwd),W(iwtcos),W(iwp))
   END SELECT
-  400  ipstor = W(iww1)
+  400  ipstor = INT( W(iww1) )
   irev = 2
   IF ( Nperod==4 ) GOTO 700
   500 CONTINUE
