+++ fnlib/dpoch1.f90
@@ -101,7 +101,7 @@ REAL(8) FUNCTION DPOCH1(A,X)
     bp = A
     IF ( A<(-0.5D0) ) bp = 1.0D0 - A - X
     incr = 0
-    IF ( bp<10.0D0 ) incr = 11.0D0 - bp
+    IF ( bp<10.0D0 ) incr = 11 - INT( bp )
     b = bp + incr
     !
     var = b + 0.5D0*(X-1.0D0)
@@ -118,7 +118,7 @@ REAL(8) FUNCTION DPOCH1(A,X)
       term = var2
       poly1 = gbern(2)*term
       !
-      nterms = -0.5D0*alneps/alnvar + 1.0D0
+      nterms = INT( -0.5D0*alneps/alnvar ) + 1
       IF ( nterms>20 ) CALL XERMSG('SLATEC','DPOCH1',&
         'NTERMS IS TOO BIG, MAYBE D1MACH(3) IS BAD'&
         ,1,2)
