+++ lin/balbak.f90
@@ -89,7 +89,7 @@ SUBROUTINE BALBAK(Nm,N,Low,Igh,Scale,M,Z)
       i = ii
       IF ( i<Low.OR.i>Igh ) THEN
         IF ( i<Low ) i = Low - ii
-        k = Scale(i)
+        k = INT(Scale(i))
         IF ( k/=i ) THEN
           !
           DO j = 1, M
