+++ pchip/chfcm.f90
@@ -109,7 +109,7 @@ INTEGER FUNCTION CHFCM(D1,D2,Delta)
   !
   IF ( Delta/=zero ) THEN
     !        DATA IS NOT CONSTANT -- PICK UP SIGN.
-    itrue = SIGN(one,Delta)
+    itrue = INT( SIGN(one,Delta) )
     a = D1/Delta
     b = D2/Delta
     IF ( (a<zero).OR.(b<zero) ) THEN
